+++ RELEASE_NOTES.md
@@ -12,6 +12,12 @@
   [#38](https://github.com/xmlunit/xmlunit.net/issues/38). And neither
   of the methods could deal with `XmlSignificantWhitespace` at all.
 
+* add `XmlWhitespaceStrippedSource` that only trims characters that
+  are considered whitespace by the [XML
+  Specification](https://www.w3.org/TR/xml11/#NT-S) from textual
+  content.
+  Issue [xmlunit/#260](https://github.com/xmlunit/xmlunit/issues/260).
+
 ## XMLUnit.NET 2.9.2 - /Released 2023-03-16/
 
 * added `NodeFilters#SatisfiesAll` and `SatifiesAny` methods to make